ai-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
scikit-learn
numpy
openai
langchain
python-multipart
